+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Glamour Dresser Alert
+
+- **Added:** A new config option "Ignore items that are already present in the armoire", useful if you want to collect duplicates.
+
 ### Portrait Helper
 
 Removed SixLabors.ImageSharp as dependency (due to new licensing requirements) and rewrote the code to use Windows Imaging Component (WIC) instead.